refactor: Complete legacy code cleanup and DDD architecture finalization
- Remove all legacy app/, dom/, and infra/ modules completely
- Remove deprecated role-based auth system in favor of structured permissions
- Clean up admin frontend removing role components and files
- Consolidate authentication flow with pure DDD implementation
- Remove legacy DTOs and use cases in favor of CQRS command/query handlers
- Finalize infrastructure layer with proper adapter patterns
- Complete migration from old infra/ to new infrastructure/ structure

+++ apps/backend/src/domain/session_management/value_objects/device_info.rs
@@ -1,9 +1,8 @@
-// Device Info Value Object
+use std::collections::HashMap;
+use chrono::{DateTime, Utc};// Device Info Value Object
 // Tracks device information for session security and analytics
 
-use chrono::{DateTime, Utc};
 use serde::{Serialize, Deserialize};
-use std::collections::HashMap;
 
 /// Device information associated with a session
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -155,7 +154,7 @@ impl DeviceInfo {
             score += 5.0;
         }
         
-        self.trust_score = score.clamp(0.0, 100.0);
+        self.trust_score = (score as f64).clamp(0.0, 100.0);
     }
 }
